# Improve performance of some Time & Rational methods